requirements.txt
-i https://pypi.org/simple
anyio==4.0.0; python_version >= '3.8'
ariadne==0.21
arrow==1.3.0; python_version >= '3.8'
blinker==1.7.0; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.1.3; python_version < '3.11'
flask==3.0.0; python_version >= '3.8'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pytest==7.4.3; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.23; python_version >= '3.7'
starlette==0.32.0.post1; python_version >= '3.8'
tomli==2.0.1; python_version < '3.11'
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
werkzeug==3.0.1; python_version >= '3.8'
wheel==0.41.3; python_version >= '3.7'
